-Created a custom hook to get routes if logged in.
